requirements.txt
click==8.0.3
dataclasses==0.6
Flask==2.0.2
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
py==1.11.0
pytest==3.0.6
typing_extensions==4.0.1
Werkzeug==2.0.2
zipp==3.6.0
azure-functions